meson: rename libbasic to libbasic_static
Our variables for internal libraries are named 'libfoo' for the shared lib
variant, and 'libfoo_static' for the static lib variant. The only exception was
libbasic, because we didn't have a shared variant for it. But let's rename it
for consitency. This makes the build config easier to understand.

see currently unmerged https://github.com/systemd/systemd/pull/33599

RHEL-only workaround

Resolves: RHEL-46020