varlink: add new calls for server-side user record filtering to varlink IDL + to spec
This is preparation for adding server side filtering to the userdb
logic: it adds some fields for this to the userdb varlink API. This only
adds the IDL for it, no client will use it for now, no server implement
it. That's added in later commits.

(cherry picked from commit 1ff1e0e01b8cddea89aad88671069527b981a9a2)

Related: RHEL-143036