journal: store counts, not byte sizes, in table size constants
It's easier to think about the size in "objects", not bytes. Let's convert to
bytes at the last moment.

Also drop some of the pointless size suffixes. In general, it's the size of the
variable that matters, not the constant that is written to it.

No functional change.

(cherry picked from commit fbbcd0edefd2921b117eebfeab8eefc3d26bd0c2)

Resolves: RHEL-106795