Enable KEY_PERFORMANCE key present on Linux 6.17
Note, this change does not require the kernel running on the host is
equal or newer than 6.17. But systemd-udevd needs to be built with the
kernel headers with KEY_PERFORMANCE, and the relevant kernel header is
already updated by the previous commit.

(cherry picked from commit d5f65056ee5dea588c168ce64c0dbbc0344d60c8)

Resolves: RHEL-72701