hwdb: fix unstable button triggering on Mipad 2 under GNOME (#40071)
Change unknown key mappings to reserved.

A KEY_RESERVED button is marked as reserved and passed to the kernel. This will stop the kernel from passing this keystroke event to user space.

If unknown, the key is set to KEY_UNKNOWN and the event is passed to user space, which we want to avoid.

(cherry picked from commit f9425092c62461a7164f4253f64929603d26af52)

Resolves: RHEL-72702